backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
razorpay==1.4.2
celery==5.3.4
redis==5.0.1
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
whitenoise==6.6.0
stripe==7.7.0